# perf: make debug a constant set using a build flag