--- a/gnovm/pkg/gnolang/debug_false.go
+++ b/gnovm/pkg/gnolang/debug_false.go
@@ -1,0 +1,5 @@
+//go:build !debug
+
+package gnolang
+
+const debug debugging = false