--- a/gnovm/pkg/gnolang/debug_true.go
+++ b/gnovm/pkg/gnolang/debug_true.go
@@ -1,0 +1,5 @@
+//go:build debug
+
+package gnolang
+
+const debug debugging = true